requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
autopep8==1.6.0
    # via -r requirements.in
azure-functions==1.11.2
    # via -r requirements.in
click==8.0.3
    # via
    #   -r requirements.in
    #   flask
colorama==0.4.4
    # via -r requirements.in
flask==2.0.2
    # via -r requirements.in
itsdangerous==2.0.1
    # via
    #   -r requirements.in
    #   flask
jinja2==3.0.3
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.0.1
    # via
    #   -r requirements.in
    #   jinja2
pycodestyle==2.8.0
    # via
    #   -r requirements.in
    #   autopep8
python-decouple==3.1
    # via -r requirements.in
regex==2022.7.25
    # via -r requirements.in
secure-smtplib==0.1.1
    # via -r requirements.in
toml==0.10.2
    # via
    #   -r requirements.in
    #   autopep8
werkzeug==2.0.2
    # via
    #   -r requirements.in
    #   flask
